requirements.txt
aiohttp==3.8.4
aiofiles==23.1.0
GitPython==3.1.31
chromadb==0.3.26
langchain==0.0.202
transformers==4.30.2
accelerate==0.20.3
python-dotenv==1.0.0
unstructured==0.7.3
tqdm==4.65.0
chainlit==0.4.0
wrapt==1.15.0
databricks-sdk==0.1.9
PyMuPDF==1.22.3
fastapi==0.96.0
pydantic==1.10.8
databricks-sql-connector==2.5.2
sentence_transformers==2.2.2
tabulate==0.9.0
tiktoken==0.4.0